# docs/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "OGGAQ Backend API",
    "version": "0.1.0",
    "description": "Uygulanmış Identity/Tenancy uçları. Soru bankası, sınav, import ve offline uçları henüz yoktur."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Yerel geliştirme"
    }
  ],
...
  "security": [
    {
      "CookieAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/mobile-tokens": {
      "post": {
        "operationId": "mobileToken",
        "summary": "Cihaz bazlı süreli mobil token",
        "description": "Aynı device_id önceki tokenı iptal eder. Süre üst sınırı yedi gündür; aktif hesap gerekir.",
        "tags": [
          "Identity"
        ],
        "security": [],
        "responses": {
          "201": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileTokenResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileTokenInput"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "forgotPassword",
        "summary": "Parola sıfırlama talebi",
        "description": "Hesabın bulunup bulunmadığını açıklamayan ortak yanıt.",
        "tags": [
          "Identity"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "summary": "Parolayı sıfırlama ve oturumları iptal etme",
        "description": "Parolayı sıfırlama ve oturumları iptal etme",
        "tags": [
          "Identity"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/email-verification-notification": {
      "post": {
        "operationId": "resendVerification",
        "summary": "Doğrulama e-postasını tekrar gönderme",
        "description": "Doğrulama e-postasını tekrar gönderme",
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/auth/verify-email/{id}/{hash}": {
      "get": {
        "operationId": "verifyEmail",
        "summary": "İmzalı bağlantıyla e-posta doğrulama",
        "description": "Giriş yapmış aynı kullanıcı, geçerli signature ve expires gerektirir.",
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "hash",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{40}$"
            }
          },
          {
            "in": "query",
            "name": "expires",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "signature",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/tenants": {
      "get": {
        "operationId": "listTenants",
        "summary": "Platform yöneticisi kurum listesi",
        "description": "Platform yöneticisi kurum listesi",
        "tags": [
          "Tenancy"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string",
              "maxLength": 2048
            }
          }
        ]
      },
      "post": {
        "operationId": "createTenant",
        "summary": "Platform yöneticisi kurum oluşturma",
        "description": "Platform yöneticisi kurum oluşturma",
        "tags": [
          "Tenancy"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantInput"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenants/{tenant}/memberships/{membership}": {
      "get": {
        "operationId": "membership",
        "summary": "Kurum içinden tek üyelik okuma",
        "description": "Kurum içinden tek üyelik okuma",
        "tags": [
          "Tenancy"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "tenant",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "membership",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/tenants/{tenant}/memberships/{membership}": {
      "delete": {
        "operationId": "revokeMembership",
        "summary": "Platform yöneticisi üyelik iptali",
        "description": "Üyeliği ve geçmiş bağlamını silmez; sonraki erişimi engeller. Tekrar aynı talep güvenlidir.",
        "tags": [
          "Tenancy"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Başarılı işlem.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "tenant",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "membership",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oggaq_session",
        "description": "Web oturumu; mutasyonlar CSRF koruması kullanır."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Sonlu Sanctum token; api:access ability gerekir."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "email_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "platform_role": {
            "type": "string",
            "enum": [
              "student",
              "platform_admin"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "email_verified_at",
          "platform_role"
        ]
      },
      "Tenant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "status"
        ]
      },
      "MemberUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
...
      "Membership": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "student",
              "company_admin"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "revoked"
            ]
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/MemberUser"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "role",
          "status",
          "joined_at",
          "revoked_at",
          "user"
        ]
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "request_id"
        ]
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "MobileTokenInput": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LoginInput"
          },
          {
            "type": "object",
            "properties": {
              "device_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "device_id"
            ]
          }
        ]
      },
      "ForgotInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "token": {
            "type": "string",
            "maxLength": 256
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128,
            "description": "Büyük/küçük harf, rakam ve sembol gerekir."
          },
          "password_confirmation": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "token",
          "password",
          "password_confirmation"
        ]
      },
...
      "TenantInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150
          },
          "slug": {
            "type": "string",
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
...
      "MobileToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Yalnız bu yanıtla verilir; güvenli cihaz deposuna yazılır."
          },
          "token_type": {
            "type": "string",
            "const": "Bearer"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "token",
          "token_type",
          "expires_at",
          "user"
        ]
      },
...
      "TenantResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Tenant"
          }
        },
        "required": [
          "data"
        ]
      },
      "MembershipResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Membership"
          }
        },
        "required": [
          "data"
        ]
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Message"
          }
        },
        "required": [
          "data"
        ]
      },
      "MobileTokenResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MobileToken"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "TenantList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tenant"
            }
          },
          "links": {
            "type": "object"
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "links",
          "meta"
        ]
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Türkçe problem ayrıntısı. 429 yanıtında Retry-After bulunur.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  }
}